+++ coWeave/View Controller/Users/UserTableViewController.swift
@@ -187,45 +187,53 @@ class UserTableViewController: UITableViewController {
         if editingStyle == .delete {
             // Fetch Record
             let record = self.fetchedResultsController.object(at: indexPath) as User
-            // Create the alert controller
-            let alertController = UIAlertController(title: "Supprimer", message: "Voulez-vous vraiment supprimer \(record.name!)? \n\n Vous ne pourrez plus rétablir ces données!", preferredStyle: .alert)
-            let deleteAction = UIAlertAction(title: "Supprimer", style: UIAlertActionStyle.destructive) {
-                UIAlertAction in
-                NSLog("Supprimer Pressed")
+            if (record.documents!.count > 0) {
+                // Create the alert controller
+                let alertController = UIAlertController(title: "Supprimer", message: "Veuillez d'abord supprimer les documents de cet utilisateur", preferredStyle: .alert)
+                let cancelAction = UIAlertAction(title: "Fermer", style: UIAlertActionStyle.cancel) {
+                    UIAlertAction in
+                    NSLog("Cancel Pressed")
+                }
+                alertController.addAction(cancelAction)
                 
-                // Delete Record
-                self.managedObjectContext.delete(record)
-                do {
-                    try self.fetchedResultsController.performFetch()
-                } catch {
-                    let fetchError = error as NSError
-                    print("\(fetchError), \(fetchError.userInfo)")
+                // Present the controller
+                self.present(alertController, animated: true, completion: nil)
+                
+            } else {
+                // Create the alert controller
+                let alertController = UIAlertController(title: "Supprimer", message: "Voulez-vous vraiment supprimer \(record.name!)? \n\n Vous ne pourrez plus rétablir ces données!", preferredStyle: .alert)
+                let deleteAction = UIAlertAction(title: "Supprimer", style: UIAlertActionStyle.destructive) {
+                    UIAlertAction in
+                    NSLog("Supprimer Pressed")
+                    
+                    // Delete Record
+                    self.managedObjectContext.delete(record)
+                    do {
+                        try self.fetchedResultsController.performFetch()
+                    } catch {
+                        let fetchError = error as NSError
+                        print("\(fetchError), \(fetchError.userInfo)")
+                    }
+                    do {
+                        // Save Record
+                        try self.managedObjectContext?.save()
+                    } catch {
+                        let saveError = error as NSError
+                        print("\(saveError), \(saveError.userInfo)")
+                    }
+                    self.tableView.deleteRows(at: [indexPath], with: .automatic)
                 }
-                do {
-                    // Save Record
-                    try self.managedObjectContext?.save()
-                } catch {
-                    let saveError = error as NSError
-                    print("\(saveError), \(saveError.userInfo)")
+                let cancelAction = UIAlertAction(title: "Annuler", style: UIAlertActionStyle.cancel) {
+                    UIAlertAction in
+                    NSLog("Cancel Pressed")
                 }
-                do {
-                    try self.fetchedResultsController.performFetch()
-                } catch {
-                    let fetchError = error as NSError
-                    print("\(fetchError), \(fetchError.userInfo)")
-                }
-                self.tableView.deleteRows(at: [indexPath], with: .automatic)
+                
+                alertController.addAction(deleteAction)
+                alertController.addAction(cancelAction)
+                
+                // Present the controller
+                self.present(alertController, animated: true, completion: nil)
             }
-            let cancelAction = UIAlertAction(title: "Annuler", style: UIAlertActionStyle.cancel) {
-                UIAlertAction in
-                NSLog("Cancel Pressed")
-            }
-            
-            alertController.addAction(deleteAction)
-            alertController.addAction(cancelAction)
-            
-            // Present the controller
-            self.present(alertController, animated: true, completion: nil)
         }
     }
